+++ electron/utils/storage.js
@@ -14,6 +14,8 @@ const store = new Store({
     colorMode: { type: 'string', default: 'light' },
     registered_nodes: { type: 'object' },
     registered_network_nodes: { type: 'object' },
+    jobs: { type: 'array', default: [] },
+    jobIdMap: { type: 'object', default: {} },
   },
 })